@@ -1,5 +1,5 @@
 module github.com/halimath/calc
 
-go 1.21
+go 1.22
 
 require github.com/halimath/expect v0.5.0 // indirect
